+++ source/edit.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <script src="./assets/scripts/edit.js"></script>
-    <script src="./assets/scripts/redirect.js"></script>
+    <script type="module" src="./assets/scripts/edit.js"></script>
+    <script type="module" src="./assets/scripts/redirect.js"></script>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
