+++ source/index.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <script src="./assets/scripts/upload.js"></script>
-    <script src="./assets/scripts/display.js"></script>
-    <script src="./assets/scripts/redirect.js"></script>
+    <script type="module" src="./assets/scripts/upload.js"></script>
+    <script type="module" src="./assets/scripts/display.js"></script>
+    <script type="module" src="./assets/scripts/redirect.js"></script>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
